Split perf tests into separate files
