+++ apps/performance-tests/src/FlatHierarchy.test.ts
@@ -1,0 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+import { runHierarchyTest } from "./util/TestUtilities";
+
+runHierarchyTest({ testName: "flat 50k elements list", iModelName: "50k elements" });
